ENH: reorganizing from preprocessing to imageoperations
